# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: InterviewOS API
  version: 1.0.0
  summary: "REST API for InterviewOS — the interview-prep learning operating system (GA track: Backend SDE3)."
  description: |
    REST API under `/api/v1`. JWT bearer access tokens (short-lived) + rotating
    refresh tokens. All field names and enum values are consistent with
    `04-DATABASE-SCHEMA.md` and `05-API-CONTRACTS.md`.
  contact:
    name: InterviewOS Engineering
    email: eng@interviewos.app
  license:
    name: Proprietary
servers:
  - url: https://api.interviewos.app/api/v1
    description: Production
  - url: https://staging-api.interviewos.app/api/v1
    description: Staging
  - url: http://localhost:8080/api/v1
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /profile:
    get:
      tags: [Profile]
      summary: Get my intake profile
      operationId: getProfile
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Profile]
      summary: Create or update intake profile
      operationId: upsertProfile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserProfileUpsert' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
        '422': { $ref: '#/components/responses/ValidationError' }

  # ---------------- CURRICULUM ----------------
...
  /resume/profile:
    get:
      tags: [Resume]
      summary: Get my resume profile
      operationId: getResumeProfile
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ResumeProfile' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Resume]
      summary: Create or update my resume profile
      operationId: upsertResumeProfile
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ResumeProfileUpsert' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ResumeProfile' } } } }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Resume]
      summary: Delete my resume profile
      operationId: deleteResumeProfile
      responses:
        '204': { description: No Content (profile deleted) }
        '404': { $ref: '#/components/responses/NotFound' }
  /resume/projects:
    get:
      tags: [Resume]
      summary: List my resume projects
      operationId: listResumeProjects
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/ResumeProject' } }
    post:
      tags: [Resume]
      summary: Add a resume project
      operationId: createResumeProject
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ResumeProjectUpsert' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/ResumeProject' } } } }
        '422': { $ref: '#/components/responses/ValidationError' }
  /resume/projects/{projectId}:
    put:
      tags: [Resume]
      summary: Update a resume project
      operationId: updateResumeProject
      parameters: [ { name: projectId, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ResumeProjectUpsert' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ResumeProject' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Resume]
      summary: Delete a resume project
      operationId: deleteResumeProject
      parameters: [ { name: projectId, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses:
        '204': { description: No Content }
        '404': { $ref: '#/components/responses/NotFound' }
  /resume/score:
    post:
      tags: [Resume, AI]
      summary: Score resume (ATS + AI review)
      operationId: scoreResume
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ResumeScoreResponse' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/AIUnavailable' }
  /resume/file:
    post:
      tags: [Resume]
      summary: Upload (or replace) my resume file
      operationId: uploadResumeFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary, description: PDF or DOCX, max 5MB }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/ResumeFileMeta' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '422': { $ref: '#/components/responses/ValidationError' }
    get:
      tags: [Resume]
      summary: Download my resume file
      operationId: downloadResumeFile
      responses:
        '200':
          description: The resume file bytes (Content-Disposition attachment).
          content:
            application/pdf: { schema: { type: string, format: binary } }
            application/vnd.openxmlformats-officedocument.wordprocessingml.document: { schema: { type: string, format: binary } }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Resume]
      summary: Delete my resume file
      operationId: deleteResumeFile
      responses:
        '204': { description: No Content }
        '404': { $ref: '#/components/responses/NotFound' }
  /resume/file/meta:
    get:
      tags: [Resume]
      summary: Get my resume file metadata
      operationId: getResumeFileMeta
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ResumeFileMeta' } } } }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------- MOCK ----------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Short-lived access token in `Authorization: Bearer <token>`."
    refreshAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Rotating refresh token (HttpOnly cookie; may also be sent in body).
...
  responses:
    BadRequest:
      description: Malformed request
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
    ValidationError:
      description: Validation failed
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
    AIUnavailable:
      description: AI provider unavailable (deterministic fallback may be returned)
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
  schemas:
...
    ConfidenceLevel: { type: integer, minimum: 1, maximum: 5 }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: Stable machine code, e.g. VALIDATION_ERROR }
            message: { type: string }
            request_id: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
...
    UserProfile:
      type: object
      required: [id, user_id, track_id, target_role, hours_per_week, start_date, target_weeks]
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        track_id: { type: string, format: uuid }
        years_experience: { type: number }
        target_company_id: { type: [string, 'null'], format: uuid }
        target_role: { type: string }
        target_level: { type: [string, 'null'] }
        hours_per_week: { type: integer }
        start_date: { type: string, format: date }
        target_weeks: { type: integer }
        pillar_strengths:
          type: object
          additionalProperties: { $ref: '#/components/schemas/ConfidenceLevel' }
        timezone: { type: string }
        onboarding_completed_at: { type: [string, 'null'], format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    UserProfileUpsert:
      type: object
      required: [track_id, target_role, hours_per_week, start_date]
      properties:
        track_id: { type: string, format: uuid }
        years_experience: { type: number, minimum: 0 }
        target_company_id: { type: [string, 'null'], format: uuid }
        target_role: { type: string }
        target_level: { type: [string, 'null'] }
        hours_per_week: { type: integer, minimum: 1, maximum: 80 }
        start_date: { type: string, format: date }
        target_weeks: { type: integer, minimum: 1, maximum: 52, default: 12 }
        pillar_strengths:
          type: object
          additionalProperties: { $ref: '#/components/schemas/ConfidenceLevel' }
        timezone: { type: string, default: UTC }
        intake_answers: { type: object, additionalProperties: true }

    # ---- Curriculum ----
...
    ResumeProfile:
      type: object
      required: [id, user_id]
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        headline: { type: [string, 'null'] }
        summary: { type: [string, 'null'] }
        years_experience: { type: [number, 'null'] }
        skills: { type: array, items: { type: string } }
        target_keywords: { type: array, items: { type: string } }
        ats_score: { type: [number, 'null'] }
        last_scored_at: { type: [string, 'null'], format: date-time }
        projects: { type: array, items: { $ref: '#/components/schemas/ResumeProject' } }
    ResumeProfileUpsert:
      type: object
      properties:
        headline: { type: string }
        summary: { type: string }
        years_experience: { type: number, minimum: 0 }
        skills: { type: array, items: { type: string } }
        target_keywords: { type: array, items: { type: string } }
    ResumeProject:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        resume_profile_id: { type: string, format: uuid }
        name: { type: string }
        role: { type: [string, 'null'] }
        description: { type: [string, 'null'] }
        impact: { type: [string, 'null'] }
        metrics: { type: array, items: { type: string } }
        tech_stack: { type: array, items: { type: string } }
        start_date: { type: [string, 'null'], format: date }
        end_date: { type: [string, 'null'], format: date }
        sort_order: { type: integer }
    ResumeProjectUpsert:
      type: object
      required: [name]
      properties:
        name: { type: string }
        role: { type: string }
        description: { type: string }
        impact: { type: string }
        metrics: { type: array, items: { type: string } }
        tech_stack: { type: array, items: { type: string } }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date }
        sort_order: { type: integer }
    ResumeScoreResponse:
      type: object
      required: [ats_score]
      properties:
        ats_score: { type: number }
        keyword_matches: { type: array, items: { type: string } }
        missing_keywords: { type: array, items: { type: string } }
        suggestions: { type: array, items: { type: string } }
        used_fallback: { type: boolean }
    ResumeFileMeta:
      type: object
      required: [file_name, content_type, size_bytes, uploaded_at]
      properties:
        file_name: { type: string }
        content_type: { type: string }
        size_bytes: { type: integer }
        uploaded_at: { type: string, format: date-time }

    # ---- Mock ----
...